Travel Manger
In this project, a reservation application has been developed with a user interface design, without using a database. The companies, vehicles, and routes mentioned in the report have been defined. Users can add a desired company and vehicle to the application, and the daily profit of the companies can be viewed. Additionally, users can select departure and arrival points from the given cities, search for available routes, and purchase tickets.